--- src/main/java/core/context/ApplicationContext.java
@@ -1,5 +1,0 @@
-package core.context;
-
-public interface ApplicationContext {
-
-}
