+++ src/main/java/core/di/context/ApplicationContext.java
@@ -1,0 +1,9 @@
+package core.di.context;
+
+import java.lang.annotation.Annotation;
+
+public interface ApplicationContext {
+    Object[] getBeans(Class<? extends Annotation> controllerClass);
+
+    <T>T getBean(Class<T> clazz);
+}
