+++ src/main/java/next/config/NextConfiguration.java
@@ -12,13 +12,18 @@ import org.apache.commons.dbcp2.BasicDataSource;
 @ComponentScan(basePackages = { "next", "core" })
 public class NextConfiguration {
 
+    private static final String DB_DRIVER = "org.h2.Driver";
+    private static final String DB_URL = "jdbc:h2:mem://localhost/~/jwp-jdbc;DB_CLOSE_DELAY=-1";
+    private static final String DB_USERNAME = "sa";
+    private static final String DB_PW = "";
+
     @Bean
     public DataSource dataSource() {
         BasicDataSource ds = new BasicDataSource();
-        ds.setDriverClassName("org.h2.Driver");
-        ds.setUrl("jdbc:h2:~/jwp-framework;MVCC=TRUE;DB_CLOSE_ON_EXIT=FALSE");
-        ds.setUsername("sa");
-        ds.setPassword("");
+        ds.setDriverClassName(DB_DRIVER);
+        ds.setUrl(DB_URL);
+        ds.setUsername(DB_USERNAME);
+        ds.setPassword(DB_PW);
         return ds;
     }
